kaaj kori

+++ Assets/Scripts/Globals.cs
@@ -4,6 +4,6 @@ public class Globals
     {
         public const float MAX_HILL_HEIGHT = 5.0f;
         public const float WIDTH = 5.0f;
-        public const float SEGMENT_COUNT = 1;
+        public const int SEGMENT_COUNT = 1;
     }
 }
